requirements.txt
# =============================================================================
# E-Commerce Data Pipeline - Python Dependencies
# =============================================================================

# Core
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Apache Kafka
kafka-python==2.0.2
confluent-kafka==2.3.0

# Apache Spark
pyspark==3.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1

# AWS/MinIO
boto3==1.33.0
minio==7.2.0

# Apache Airflow
apache-airflow==2.8.0
apache-airflow-providers-postgres==5.7.0
apache-airflow-providers-apache-spark==4.5.0
apache-airflow-providers-apache-kafka==1.3.0

# HTTP/API
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Data Validation
great-expectations==0.18.0
pandera==0.17.2

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-asyncio==0.23.0

# Linting
flake8==6.1.0
black==23.11.0
isort==5.12.0
mypy==1.7.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
tqdm==4.66.1
click==8.1.7
rich==13.7.0
tenacity==8.2.3